+++ src/utils/transforms.py
@@ -134,7 +134,8 @@ tumor_seg_transform = {
             pixdim=(1.0, 1.0, 1.0),
             mode=("bilinear", "nearest"),
         ),
-        RandSpatialCropd(keys=["image", "label"], roi_size=[224, 224, 144], random_size=False),
+        # RandSpatialCropd(keys=["image", "label"], roi_size=[224, 224, 144], random_size=False),
+        RandSpatialCropd(keys=["image", "label"], roi_size=[64, 64, 64], random_size=False),
         RandFlipd(keys=["image", "label"], prob=0.5, spatial_axis=0),
         RandFlipd(keys=["image", "label"], prob=0.5, spatial_axis=1),
         RandFlipd(keys=["image", "label"], prob=0.5, spatial_axis=2),
