+++ training_notes.txt
@@ -1,3 +1,12 @@
+TODO
+----
+SLURM_JOB_ID = 
+RUN_ID = 22
+QR_REGRESSION = True
+DO_MASK = True
+MAX_EPOCHS = 6000
+TRAIN_DATA_SIZE = 200
+
 COMPLETED
 ----
 SLURM_JOB_ID = 27571733
